Rename RayBooks to Palace.
