+++ .ci-local/deploy-firebase-apps.conf
@@ -1,4 +1,4 @@
-simplified-app-raybooks
+simplified-app-palace
 #
 # The non-empty lines of this file that do not begin with '#' are interpreted as the names
 # of submodules that will have builds deployed with Firebase.
