+++ .ci-local/deploy-git-binary-version-file.conf
@@ -1,4 +1,4 @@
-simplified-app-raybooks/version.properties
+simplified-app-palace/version.properties
 
 #
 # The first line of this file specifies the name of the file that is expected to contain a
