+++ settings.gradle
@@ -16,7 +16,7 @@ include ':simplified-analytics-circulation'
 include ':simplified-android-ktx'
 include ':simplified-announcements'
 //include ':simplified-app-openebooks'
-include ':simplified-app-raybooks'
+include ':simplified-app-palace'
 //include ':simplified-app-simplye'
 include ':simplified-app-vanilla'
 include ':simplified-books-api'
