+++ simplified-app-palace/firebase-apk.conf
@@ -1,4 +1,4 @@
-build/outputs/apk/release/raybooks-release.apk
+build/outputs/apk/release/palace-release.apk
 #
 # The first line of this file is the APK that will be deployed to Firebase.
 #
