+++ simplified-app-palace/firebase-app-id.conf
