+++ simplified-app-palace/firebase-groups.conf
