+++ simplified-app-palace/gradle.properties
@@ -1,6 +1,6 @@
-POM_ARTIFACT_ID=org.lyrasis.raybooks
-POM_DESCRIPTION=RayBooks
-POM_NAME=org.lyrasis.raybooks
+POM_ARTIFACT_ID=org.thepalaceproject.palace
+POM_DESCRIPTION=Palace
+POM_NAME=org.thepalaceproject.palace
 POM_PACKAGING=apk
 
 # Skip this project unless the property `org.librarysimplified.drm.enabled=true`
