+++ simplified-app-palace/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
-  package="org.lyrasis.raybooks">
+  package="org.thepalaceproject.palace">
 
   <uses-permission android:name="android.permission.INTERNET" />
   <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -15,7 +15,7 @@
     android:contentDescription="@string/app_name"
     android:icon="@mipmap/ic_launcher"
     android:label="@string/app_name"
-    android:theme="@style/RayBooksTheme_NoActionBar"
+    android:theme="@style/PalaceTheme_NoActionBar"
     android:usesCleartextTraffic="true"
     tools:replace="android:allowBackup">
 
@@ -34,7 +34,7 @@
         <action android:name="android.intent.action.VIEW" />
         <category android:name="android.intent.category.DEFAULT" />
         <category android:name="android.intent.category.BROWSABLE" />
-        <data android:scheme="raybooks-oauth" />
+        <data android:scheme="palace-oauth" />
       </intent-filter>
     </activity>
 
@@ -72,7 +72,7 @@
       android:exported="false"
       android:parentActivityName="org.nypl.simplified.main.MainActivity"
       android:screenOrientation="portrait"
-      android:theme="@style/RayBooksTheme_NoActionBar" />
+      android:theme="@style/PalaceTheme_NoActionBar" />
 
     <activity
       android:name="org.nypl.simplified.viewer.pdf.PdfReaderActivity"
