+++ simplified-app-palace/src/main/assets/about.html
