+++ simplified-app-palace/src/main/assets/eula.html
