+++ simplified-app-palace/src/main/assets/software-licenses.html
