+++ simplified-app-palace/src/main/java/org/thepalaceproject/palace/BrandingSplashService.kt
@@ -1,4 +1,4 @@
-package org.lyrasis.raybooks
+package org.thepalaceproject.palace
 
 import org.nypl.simplified.ui.branding.BrandingSplashServiceType
