+++ simplified-app-palace/src/main/java/org/thepalaceproject/palace/BrandingThemeService.kt
@@ -1,4 +1,4 @@
-package org.lyrasis.raybooks
+package org.thepalaceproject.palace
 
 import org.nypl.simplified.ui.branding.BrandingThemeOverrideServiceType
 import org.nypl.simplified.ui.theme.ThemeValue
@@ -6,12 +6,12 @@ import org.nypl.simplified.ui.theme.ThemeValue
 class BrandingThemeService : BrandingThemeOverrideServiceType {
   override fun overrideTheme(): ThemeValue {
     return ThemeValue(
-      name = "raybooks",
+      name = "palace",
       color = R.color.colorPrimary,
       colorLight = R.color.colorPrimaryLight,
       colorDark = R.color.colorPrimaryDark,
-      themeWithActionBar = R.style.RayBooksTheme_ActionBar,
-      themeWithNoActionBar = R.style.RayBooksTheme_NoActionBar
+      themeWithActionBar = R.style.PalaceTheme_ActionBar,
+      themeWithNoActionBar = R.style.PalaceTheme_NoActionBar
     )
   }
 }
