+++ simplified-app-palace/src/main/res/drawable/main_icon.png
