+++ simplified-app-palace/src/main/res/mipmap-hdpi/ic_launcher.png
