+++ simplified-app-palace/src/main/res/mipmap-hdpi/ic_launcher_foreground.png
