+++ simplified-app-palace/src/main/res/mipmap-xhdpi/ic_launcher_foreground.png
