+++ simplified-app-palace/src/main/res/mipmap-xxhdpi/ic_launcher.png
