+++ simplified-app-palace/src/main/res/mipmap-xxhdpi/ic_launcher_foreground.png
