+++ simplified-app-palace/src/main/res/mipmap-xxxhdpi/ic_launcher.png
