+++ simplified-app-palace/src/main/res/values/colors.xml
