+++ simplified-app-palace/src/main/res/values/features.xml
@@ -2,5 +2,6 @@
 
 <!-- Build-time feature flags. -->
 <resources>
+  <!-- TODO: Change to Palace when the certificate is updated -->
   <string name="featureAdobeDRMPackageOverride" translatable="false">Raybooks</string>
 </resources>
