+++ simplified-app-palace/src/main/res/values/ic_launcher_background.xml
