+++ simplified-app-palace/src/main/res/values/styles.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-  <style name="RayBooksTheme_NoActionBar" parent="SimplifiedTheme.NoActionBar.Base">
+  <style name="PalaceTheme_NoActionBar" parent="SimplifiedTheme.NoActionBar.Base">
     <item name="simplifiedColorPrimaryLight">@color/colorPrimaryLight</item>
     <item name="simplifiedColorPrimary">@color/colorPrimary</item>
     <item name="simplifiedColorPrimaryDark">@color/colorPrimaryDark</item>
   </style>
 
-  <style name="RayBooksTheme_ActionBar" parent="SimplifiedTheme.ActionBar.Base">
+  <style name="PalaceTheme_ActionBar" parent="SimplifiedTheme.ActionBar.Base">
     <item name="simplifiedColorPrimaryLight">@color/colorPrimaryLight</item>
     <item name="simplifiedColorPrimary">@color/colorPrimary</item>
     <item name="simplifiedColorPrimaryDark">@color/colorPrimaryDark</item>
