+++ simplified-app-palace/src/main/resources/META-INF/services/org.librarysimplified.documents.DocumentConfigurationServiceType
@@ -1,0 +1,1 @@
+org.thepalaceproject.palace.PalaceDocumentStoreConfiguration
