+++ simplified-app-palace/src/main/resources/META-INF/services/org.nypl.simplified.accounts.api.AccountProviderFallbackType
@@ -1,0 +1,1 @@
+org.thepalaceproject.palace.PalaceAccountFallback
