+++ simplified-app-palace/src/main/resources/META-INF/services/org.nypl.simplified.buildconfig.api.BuildConfigurationServiceType
@@ -1,0 +1,1 @@
+org.thepalaceproject.palace.PalaceBuildConfigurationService
