+++ simplified-app-palace/src/main/resources/META-INF/services/org.nypl.simplified.ui.branding.BrandingSplashServiceType
@@ -1,0 +1,1 @@
+org.thepalaceproject.palace.BrandingSplashService
