+++ simplified-app-palace/src/main/xcf/main_splash.xcf
