--- simplified-app-raybooks/google-services.json
@@ -1,47 +1,0 @@
-{
-  "project_info": {
-    "project_number": "287525833402",
-    "firebase_url": "https://raybooks-1ef32-default-rtdb.firebaseio.com",
-    "project_id": "raybooks-1ef32",
-    "storage_bucket": "raybooks-1ef32.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:287525833402:android:06fdf81f003d8a1c9bd083",
-        "android_client_info": {
-          "package_name": "org.lyrasis.raybooks"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "287525833402-9g3mbpq7bb68hhhalvntl9v7httiatru.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyDK69dMWpbJ0VwU2WHKVnnJHqe0igbORtY"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": [
-            {
-              "client_id": "287525833402-9g3mbpq7bb68hhhalvntl9v7httiatru.apps.googleusercontent.com",
-              "client_type": 3
-            },
-            {
-              "client_id": "287525833402-5jl0eaclkq82op1krg85c9esk6jl9je1.apps.googleusercontent.com",
-              "client_type": 2,
-              "ios_info": {
-                "bundle_id": "org.lyrasis.raybooks"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
