--- simplified-app-raybooks/src/main/resources/META-INF/services/org.librarysimplified.documents.DocumentConfigurationServiceType
@@ -1,1 +1,0 @@
-org.lyrasis.raybooks.RayBooksDocumentStoreConfiguration
