--- simplified-app-raybooks/src/main/resources/META-INF/services/org.nypl.simplified.accounts.api.AccountProviderFallbackType
@@ -1,1 +1,0 @@
-org.lyrasis.raybooks.RayBooksAccountFallback
