--- simplified-app-raybooks/src/main/resources/META-INF/services/org.nypl.simplified.buildconfig.api.BuildConfigurationServiceType
@@ -1,1 +1,0 @@
-org.lyrasis.raybooks.RayBooksBuildConfigurationService
