--- simplified-app-raybooks/src/main/resources/META-INF/services/org.nypl.simplified.ui.branding.BrandingSplashServiceType
@@ -1,1 +1,0 @@
-org.lyrasis.raybooks.BrandingSplashService
