--- simplified-app-raybooks/src/main/resources/META-INF/services/org.nypl.simplified.ui.branding.BrandingThemeOverrideServiceType
@@ -1,1 +1,0 @@
-org.lyrasis.raybooks.BrandingThemeService
